# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SyncOps AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  total_services: number;
  healthy_services: number;
  open_incidents: number;
  critical_incidents: number;
  recent_deployments: number;
  failed_deployments: number;
  avg_error_rate: number;
  avg_latency: number;
  total_logs_24h: number;
  error_logs_24h: number;
}

export interface Service {
  id: string;
  project_id: string;
  name: string;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  owner_team?: string | null;
  tier: string;
  /** @nullable */
  description?: string | null;
  created_at: string;
  updated_at: string;
}

export type ServiceInputTier = typeof ServiceInputTier[keyof typeof ServiceInputTier];


export const ServiceInputTier = {
  normal: 'normal',
  high: 'high',
  critical: 'critical',
} as const;

export interface ServiceInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  owner_team?: string | null;
  tier?: ServiceInputTier;
  /** @nullable */
  description?: string | null;
}

export type ServiceUpdateTier = typeof ServiceUpdateTier[keyof typeof ServiceUpdateTier];


export const ServiceUpdateTier = {
  normal: 'normal',
  high: 'high',
  critical: 'critical',
} as const;

export interface ServiceUpdate {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name?: string;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  owner_team?: string | null;
  tier?: ServiceUpdateTier;
  /** @nullable */
  description?: string | null;
}

export interface Metric {
  id: string;
  service_id: string;
  metric_type: string;
  value: number;
  timestamp: string;
}

export interface MetricSummary {
  avg_latency: number;
  avg_error_rate: number;
  avg_throughput: number;
  total_services: number;
}

export interface DemoGenerateResult {
  metrics_generated: number;
  services_count: number;
}

export interface LogEntry {
  id: string;
  /** @nullable */
  service_id?: string | null;
  /** @nullable */
  service_name?: string | null;
  level: string;
  message: string;
  /** @nullable */
  trace_id?: string | null;
  /** @nullable */
  metadata?: string | null;
  timestamp: string;
}

export type LogInputLevel = typeof LogInputLevel[keyof typeof LogInputLevel];


export const LogInputLevel = {
  debug: 'debug',
  info: 'info',
  warn: 'warn',
  error: 'error',
} as const;

export interface LogInput {
  /** @nullable */
  service_id?: string | null;
  level: LogInputLevel;
  message: string;
  /** @nullable */
  trace_id?: string | null;
  /** @nullable */
  metadata?: string | null;
}

export interface Incident {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  status: string;
  severity: string;
  /** @nullable */
  service_id?: string | null;
  /** @nullable */
  service_name?: string | null;
  /** @nullable */
  assigned_to?: string | null;
  created_at: string;
  updated_at: string;
  /** @nullable */
  resolved_at?: string | null;
}

export type IncidentInputSeverity = typeof IncidentInputSeverity[keyof typeof IncidentInputSeverity];


export const IncidentInputSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface IncidentInput {
  /** @minLength 1 */
  title: string;
  /** @nullable */
  description?: string | null;
  severity: IncidentInputSeverity;
  /** @nullable */
  service_id?: string | null;
  /** @nullable */
  assigned_to?: string | null;
}

export type IncidentUpdateStatus = typeof IncidentUpdateStatus[keyof typeof IncidentUpdateStatus];


export const IncidentUpdateStatus = {
  open: 'open',
  investigating: 'investigating',
  resolved: 'resolved',
} as const;

export type IncidentUpdateSeverity = typeof IncidentUpdateSeverity[keyof typeof IncidentUpdateSeverity];


export const IncidentUpdateSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface IncidentUpdate {
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: IncidentUpdateStatus;
  severity?: IncidentUpdateSeverity;
  /** @nullable */
  assigned_to?: string | null;
}

export interface Deployment {
  id: string;
  /** @nullable */
  service_id?: string | null;
  /** @nullable */
  service_name?: string | null;
  version: string;
  status: string;
  environment: string;
  /** @nullable */
  deployed_by?: string | null;
  /** @nullable */
  notes?: string | null;
  started_at: string;
  /** @nullable */
  finished_at?: string | null;
  created_at: string;
}

export type DeploymentInputEnvironment = typeof DeploymentInputEnvironment[keyof typeof DeploymentInputEnvironment];


export const DeploymentInputEnvironment = {
  production: 'production',
  staging: 'staging',
  development: 'development',
} as const;

export interface DeploymentInput {
  /** @nullable */
  service_id?: string | null;
  /** @minLength 1 */
  version: string;
  environment: DeploymentInputEnvironment;
  /** @nullable */
  deployed_by?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type DeploymentUpdateStatus = typeof DeploymentUpdateStatus[keyof typeof DeploymentUpdateStatus];


export const DeploymentUpdateStatus = {
  pending: 'pending',
  running: 'running',
  success: 'success',
  failed: 'failed',
  rolled_back: 'rolled_back',
} as const;

export interface DeploymentUpdate {
  status?: DeploymentUpdateStatus;
  /** @nullable */
  finished_at?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type ListLogsParams = {
service_id?: string;
level?: string;
search?: string;
limit?: number;
};

export type ListIncidentsParams = {
status?: string;
severity?: string;
};

export type ListDeploymentsParams = {
service_id?: string;
status?: string;
};