# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SyncOps AI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: dashboard
  - name: services
  - name: metrics
  - name: logs
  - name: incidents
  - name: deployments

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard stats summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a new service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
    put:
      operationId: updateService
      tags: [services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
    delete:
      operationId: deleteService
      tags: [services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /metrics:
    get:
      operationId: listMetrics
      tags: [metrics]
      responses:
        "200":
          description: List of metrics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Metric"

  /metrics/summary:
    get:
      operationId: getMetricsSummary
      tags: [metrics]
      responses:
        "200":
          description: Metrics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricSummary"

  /metrics/generate-demo-data:
    post:
      operationId: generateDemoMetrics
      tags: [metrics]
      responses:
        "200":
          description: Generation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoGenerateResult"

  /logs:
    get:
      operationId: listLogs
      tags: [logs]
      summary: List logs with optional filters
      parameters:
        - name: service_id
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
    post:
      operationId: ingestLog
      tags: [logs]
      summary: Ingest a log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogEntry"

  /logs/generate-demo-data:
    post:
      operationId: generateDemoLogs
      tags: [logs]
      responses:
        "200":
          description: Generation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoGenerateResult"

  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List all incidents
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
    post:
      operationId: createIncident
      tags: [incidents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"

  /incidents/{id}:
    get:
      operationId: getIncident
      tags: [incidents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "404":
          description: Not found
    patch:
      operationId: updateIncident
      tags: [incidents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "404":
          description: Not found

  /deployments:
    get:
      operationId: listDeployments
      tags: [deployments]
      parameters:
        - name: service_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
    post:
      operationId: createDeployment
      tags: [deployments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"

  /deployments/{id}:
    patch:
      operationId: updateDeployment
      tags: [deployments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "404":
          description: Not found

  /deployments/generate-demo-data:
    post:
      operationId: generateDemoDeployments
      tags: [deployments]
      responses:
        "200":
          description: Generation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoGenerateResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    DashboardSummary:
      type: object
      properties:
        total_services:
          type: integer
        healthy_services:
          type: integer
        open_incidents:
          type: integer
        critical_incidents:
          type: integer
        recent_deployments:
          type: integer
        failed_deployments:
          type: integer
        avg_error_rate:
          type: number
        avg_latency:
          type: number
        total_logs_24h:
          type: integer
        error_logs_24h:
          type: integer
      required:
        - total_services
        - healthy_services
        - open_incidents
        - critical_incidents
        - recent_deployments
        - failed_deployments
        - avg_error_rate
        - avg_latency
        - total_logs_24h
        - error_logs_24h

    Service:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        name:
          type: string
        language:
          type: ["string", "null"]
        owner_team:
          type: ["string", "null"]
        tier:
          type: string
        description:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
      required: [id, project_id, name, tier, created_at, updated_at]

    ServiceInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        language:
          type: ["string", "null"]
        owner_team:
          type: ["string", "null"]
        tier:
          type: string
          enum: [normal, high, critical]
        description:
          type: ["string", "null"]
      required: [name]

    ServiceUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        language:
          type: ["string", "null"]
        owner_team:
          type: ["string", "null"]
        tier:
          type: string
          enum: [normal, high, critical]
        description:
          type: ["string", "null"]

    Metric:
      type: object
      properties:
        id:
          type: string
        service_id:
          type: string
        metric_type:
          type: string
        value:
          type: number
        timestamp:
          type: string
      required: [id, service_id, metric_type, value, timestamp]

    MetricSummary:
      type: object
      properties:
        avg_latency:
          type: number
        avg_error_rate:
          type: number
        avg_throughput:
          type: number
        total_services:
          type: integer
      required: [avg_latency, avg_error_rate, avg_throughput, total_services]

    DemoGenerateResult:
      type: object
      properties:
        metrics_generated:
          type: integer
        services_count:
          type: integer
      required: [metrics_generated, services_count]

    LogEntry:
      type: object
      properties:
        id:
          type: string
        service_id:
          type: ["string", "null"]
        service_name:
          type: ["string", "null"]
        level:
          type: string
        message:
          type: string
        trace_id:
          type: ["string", "null"]
        metadata:
          type: ["string", "null"]
        timestamp:
          type: string
      required: [id, level, message, timestamp]

    LogInput:
      type: object
      properties:
        service_id:
          type: ["string", "null"]
        level:
          type: string
          enum: [debug, info, warn, error]
        message:
          type: string
        trace_id:
          type: ["string", "null"]
        metadata:
          type: ["string", "null"]
      required: [level, message]

    Incident:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        severity:
          type: string
        service_id:
          type: ["string", "null"]
        service_name:
          type: ["string", "null"]
        assigned_to:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
        resolved_at:
          type: ["string", "null"]
      required: [id, title, status, severity, created_at, updated_at]

    IncidentInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        severity:
          type: string
          enum: [critical, high, medium, low]
        service_id:
          type: ["string", "null"]
        assigned_to:
          type: ["string", "null"]
      required: [title, severity]

    IncidentUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, investigating, resolved]
        severity:
          type: string
          enum: [critical, high, medium, low]
        assigned_to:
          type: ["string", "null"]

    Deployment:
      type: object
      properties:
        id:
          type: string
        service_id:
          type: ["string", "null"]
        service_name:
          type: ["string", "null"]
        version:
          type: string
        status:
          type: string
        environment:
          type: string
        deployed_by:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        started_at:
          type: string
        finished_at:
          type: ["string", "null"]
        created_at:
          type: string
      required: [id, version, status, environment, started_at, created_at]

    DeploymentInput:
      type: object
      properties:
        service_id:
          type: ["string", "null"]
        version:
          type: string
          minLength: 1
        environment:
          type: string
          enum: [production, staging, development]
        deployed_by:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [version, environment]

    DeploymentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, running, success, failed, rolled_back]
        finished_at:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]