# lib/api-zod/src/generated/types/metric.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SyncOps AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Metric {
  id: string;
  service_id: string;
  metric_type: string;
  value: number;
  timestamp: string;
}